backend/requirements/base.txt
# Base
pytz==2020.1
python-slugify==4.0.0
Pillow==7.1.2
psycopg2==2.8.5 --no-binary psycopg2

# Django
django==3.0.8

# Tools
django-extensions==2.2.9

# Environment
django-environ==0.4.5

# Django REST Framework
djangorestframework == 3.11.0
django-filter==2.2.0
django-cors-headers == 3.2.1

# JWT
pyjwt==1.7.1

# Passwords security
argon2-cffi==19.2.0

# Static files
whitenoise==5.0.1

# Celery
redis==3.5.0
django-redis==4.11.0

celery==4.4.2
tornado==6.0.4
flower==0.9.4

django-material-admin